test(dataset): relocate snapshot tests to unit_tests/dataset with factory fixtures
- lift 8 helper builders into dataset conftest as factory fixtures
- drop redundant local _isolated_caches (conftest autouse covers it)
- split CI-gate tests into test_snapshot_health.py with location-robust loader
- delete 1218-line monolith tests/test_dataset_snapshot.py
